feature/calculator/common module
