+++ calculator/src/main/java/ru/neoflex/bank/calculator/exception/ErrorResponse.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import io.swagger.v3.oas.annotations.Hidden;
 import lombok.Data;
 import org.springframework.http.HttpStatus;
-import ru.neoflex.bank.calculator.util.DateTimeUtils;
+import ru.neoflex.bank.util.DateTimeUtils;
 
 import java.util.Date;
