+++ calculator/src/main/java/ru/neoflex/bank/calculator/service/AnnuityCalculatorService.java
@@ -3,8 +3,8 @@ package ru.neoflex.bank.calculator.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.neoflex.bank.calculator.configuration.RateConfig;
-import ru.neoflex.bank.calculator.model.dto.PaymentScheduleElementDto;
-import ru.neoflex.bank.calculator.util.MoneyUtils;
+import ru.neoflex.bank.model.dto.PaymentScheduleElementDto;
+import ru.neoflex.bank.util.MoneyUtils;
 import ru.neoflex.bank.logging.Logging;
 
 import java.math.BigDecimal;
