+++ common/src/main/java/ru/neoflex/bank/logging/ExceptionLogging.java
