+++ common/src/main/java/ru/neoflex/bank/logging/ExceptionLoggingAspect.java
