+++ common/src/main/java/ru/neoflex/bank/logging/Logging.java
