+++ common/src/main/java/ru/neoflex/bank/logging/LoggingAspect.java
