+++ common/src/main/java/ru/neoflex/bank/model/dto/CreditDto.java
@@ -1,36 +1,36 @@
-package ru.neoflex.bank.calculator.model.dto;
+package ru.neoflex.bank.model.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
-import ru.neoflex.bank.calculator.util.MoneyUtils;
+import ru.neoflex.bank.util.MoneyUtils;
 
 import java.math.BigDecimal;
-import java.util.UUID;
+import java.util.List;
 
 @Builder
-@Schema(description = "Кредитное предложение")
-public record LoanOfferDto(
-        @Schema(description = "Номер заявки", example = "09b84931-b217-426a-a618-4b51af2d3bb1")
-        UUID statementId,
-        @Schema(description = "Запрашиваемая сумма", example = "1000000.50")
-        BigDecimal requestedAmount,
-        @Schema(description = "Полная стоимость кредита (ПСК)", example = "331540.69")
-        BigDecimal totalAmount,
+@Schema(description = "Параметры выдаваемого кредита")
+public record CreditDto(
+        @Schema(description = "Сумма кредита", example = "1000000.50")
+        BigDecimal amount,
         @Schema(description = "Срок кредита в месяцах", example = "40")
         Integer term,
         @Schema(description = "Ежемесячный платеж", example = "17560.31")
         BigDecimal monthlyPayment,
         @Schema(description = "Ставка по кредиту в %", example = "17.4")
         BigDecimal rate,
+        @Schema(description = "Полная стоимость кредита (ПСК)", example = "331540.69")
+        BigDecimal psk,
         @Schema(description = "Включена ли страховка", example = "true")
         Boolean isInsuranceEnabled,
         @Schema(description = "Являетесь ли зарплатным клиентом", example = "false")
-        Boolean isSalaryClient
+        Boolean isSalaryClient,
+        @Schema(description = "График платежей")
+        List<PaymentScheduleElementDto> paymentSchedule
 ) {
-    public LoanOfferDto {
-        requestedAmount = MoneyUtils.moneyScale(requestedAmount);
-        totalAmount = MoneyUtils.moneyScale(totalAmount);
+    public CreditDto {
+        amount = MoneyUtils.moneyScale(amount);
         monthlyPayment = MoneyUtils.moneyScale(monthlyPayment);
         rate = MoneyUtils.rateScale(rate);
+        psk = MoneyUtils.moneyScale(psk);
     }
 }
